Colocate route lifecycle components with the modules that own them
