+++ frontend/lint/module-boundaries.mjs
@@ -191,8 +191,8 @@ const elements = [
     "frontend/src/metabase/routes-embed.tsx",
     "frontend/src/metabase/route-guards.tsx",
     "frontend/src/metabase/route-guards.unit.spec.tsx",
-    "frontend/src/metabase/route-lifecycle.tsx",
-    "frontend/src/metabase/route-lifecycle.unit.spec.tsx",
+    "frontend/src/metabase/LoadCurrentUser.tsx",
+    "frontend/src/metabase/LoadCurrentUser.unit.spec.tsx",
     "frontend/src/metabase/routes-public.tsx",
     "frontend/src/metabase/AppThemeProvider.tsx",
     "frontend/src/metabase/AppColorSchemeProvider.tsx",
