+++ frontend/src/metabase/auth/components/SsoReload/index.ts
@@ -1,0 +1,1 @@
+export * from "./SsoReload";
