+++ frontend/src/metabase/home/components/LandingPageRedirect/index.ts
@@ -1,0 +1,1 @@
+export * from "./LandingPageRedirect";
