+++ frontend/src/metabase/setup/components/RedirectIfSetup/index.ts
@@ -1,0 +1,1 @@
+export * from "./RedirectIfSetup";
